Fixes for review

+++ kernel_selector/core/cl_kernels/convolution_grad_weights_gpu_ref.cl
@@ -68,16 +68,16 @@ KERNEL(convolution_grad_weights_gpu_ref)(
                 const int input_offset_x = in_x + j * STRIDE_SIZE_X;
                 const bool zero_x = input_offset_x >= INPUT1_SIZE_X || input_offset_x < 0;
 #if BIAS_TERM
-                uint input_grad_idx = grad_split_offset + b*INPUT0_BATCH_PITCH + (uint)ofm*INPUT0_FEATURE_PITCH + j*INPUT0_X_PITCH + i*INPUT0_Y_PITCH;
+                uint input_grad_idx = grad_split_offset + b*INPUT0_BATCH_PITCH + ofm*INPUT0_FEATURE_PITCH + j*INPUT0_X_PITCH + i*INPUT0_Y_PITCH;
                 UNIT_TYPE grad = input_grad[input_grad_idx];
 #endif
                 if(!zero_x && !zero_y)
                 {
-                    uint input_idx = in_split_offset + b*INPUT1_BATCH_PITCH + (uint)ifm*INPUT1_FEATURE_PITCH + (uint)input_offset_x*INPUT1_X_PITCH + (uint)input_offset_y*INPUT1_Y_PITCH;
+                    uint input_idx = in_split_offset + b*INPUT1_BATCH_PITCH + ifm*INPUT1_FEATURE_PITCH + (uint)input_offset_x*INPUT1_X_PITCH + (uint)input_offset_y*INPUT1_Y_PITCH;
 #if BIAS_TERM
                     result = fma(input[input_idx], grad, result);
 #else
-                    uint input_grad_idx = grad_split_offset + b*INPUT1_BATCH_PITCH + (uint)ofm*INPUT0_FEATURE_PITCH + j*INPUT0_X_PITCH + i*INPUT0_Y_PITCH;
+                    uint input_grad_idx = grad_split_offset + b*INPUT1_BATCH_PITCH + ofm*INPUT0_FEATURE_PITCH + j*INPUT0_X_PITCH + i*INPUT0_Y_PITCH;
                     result = fma(input[input_idx], input_grad[input_grad_idx], result);
 #endif
                 }
